#249 - Move class gui::items::View into directory gui/views (cpp)
Completed.
